Move flatten_lists to common file
