+++ stable_baselines/ppo1/pposgd_simple.py
@@ -14,7 +14,9 @@ from stable_baselines.common.tf_util import total_episode_reward_logger
 from stable_baselines.common.policies import ActorCriticPolicy
 from stable_baselines.common.mpi_adam import MpiAdam
 from stable_baselines.common.mpi_moments import mpi_moments
-from stable_baselines.trpo_mpi.utils import traj_segment_generator, add_vtarg_and_adv, flatten_lists
+from stable_baselines.common.misc_util import flatten_lists
+from stable_baselines.trpo_mpi.utils import traj_segment_generator, add_vtarg_and_adv
+
 
 
 class PPO1(ActorCriticRLModel):
