+++ stable_baselines/trpo_mpi/trpo_mpi.py
@@ -15,7 +15,8 @@ from stable_baselines import logger
 from stable_baselines.common.mpi_adam import MpiAdam
 from stable_baselines.common.cg import conjugate_gradient
 from stable_baselines.common.policies import ActorCriticPolicy
-from stable_baselines.trpo_mpi.utils import traj_segment_generator, add_vtarg_and_adv, flatten_lists
+from stable_baselines.common.misc_util import flatten_lists
+from stable_baselines.trpo_mpi.utils import traj_segment_generator, add_vtarg_and_adv
 
 
 class TRPO(ActorCriticRLModel):
